backend/pyproject.toml
[project]
name = "flexcrm"
version = "0.1.0"
description = "This wow API for auf FlexCRM system!"
authors = [
    {name = "Ira11111",email = "mirina-0208@yandex.ru"}
]
packages = [
    { include = "crmapp" },
    { include = "crmauth" },
    { include = "flexcrm" }
]
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
    "django (==5.1.6)",
    "djangorestframework (==3.15.2)",
    "django-filter (==22.1)",
    "djangorestframework-simplejwt (==5.5.0)",
    "django-cors-headers (==4.7.0)",
    "django-phonenumber-field (==8.0.0)",
    "pillow (==11.1.0)",
    "drf-spectacular (==0.28.0)"

]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
